backend/requirements.txt
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask==2.1.2
google-auth==2.6.6
mysql-connector-python==8.0.29
requests==2.27.1
